migrate `@now/python-wsgi to `@now/python`
