--- packages/now-python-wsgi/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@now/python-wsgi",
-  "version": "0.0.1",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "execa": "^1.0.0",
-    "fs.promised": "^3.0.0",
-    "node-fetch": "^2.2.0"
-  },
-  "peerDependencies": {
-    "@now/build-utils": ">=0.0.1"
-  }
-}
